feat: shop system, loot, upgrades, combat polish
- Shop with randomized upgrade slots, pricing, cash system
- Weapon slot display (HUD + shop), weapon replacement UI
- Loot drop system with scatter, magnetism, collect VFX
- Wave pause: camera transition, shooting disabled during zoom, loot cleanup
- Hit invincibility layer-based collision ignore
- DOTween frame shake on shop item hit and cash reject
- Damage numbers, TriBullet, health bar fix, enemy null player fix

+++ Assets/Scripts/PlayerHealthUI.cs
@@ -8,12 +8,12 @@ public class PlayerHealthUI : MonoBehaviour
     [SerializeField] private TextMeshProUGUI healthText;
     [SerializeField] private Image healthBarFill;
     // Start is called once before the first execution of Update after the MonoBehaviour is created
-    public void UpdateUI(float currentHealth, float maxHealth)
+    public void UpdateUI(int currentHealth, int maxHealth)
     {
         if (healthText != null && healthBarFill != null)
         {
             healthText.text = $"{currentHealth}/{maxHealth}";
-            healthBarFill.fillAmount = currentHealth / maxHealth;
+            healthBarFill.fillAmount = (float)currentHealth / maxHealth;
         }
     }
 }
